requirements.txt
pytest==7.4.3
pytest-dependency==0.5.1
pytest-forked==1.4.0
pytest-html==1.22.0
pytest-json-report==1.2.4
pytest-metadata==2.0.1
pytest-parallel==0.0.9
pytest-rerunfailures==9.1.1
pytest-xdist==1.30.0
pytest-mock-server==0.3.0
python-dateutil==2.8.2
python-decouple==3.1


allure-pytest==2.13.2
allure-python-commons===2.13.2

requests==2.31.0
requests-oauthlib==1.3.1
beautifulsoup4==4.8.1
browserstack-local==1.2.2

selenium==4.9.1
selenium-wire==5.1.0

uritemplate==3.0.1
uritools==4.0.0
urlextract==0.14.0
urllib3==1.26.13

pandas==2.1.3